+++ designsystem/src/test/kotlin/com/natura/android/menu/MenuViewTest.kt
@@ -8,8 +8,10 @@ import androidx.appcompat.widget.AppCompatImageView
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
 import com.natura.android.R
+import com.natura.android.resources.getColorTokenFromTheme
 import kotlinx.android.synthetic.main.ds_menu_view_in_layout.view.*
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNotEquals
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -63,5 +65,26 @@ class MenuViewTest {
         val label = layout.menuItemBase.label
 
         assertEquals("Menu item to test", label)
+    }
+
+    @Test
+    fun checksLowEmphasisIsFalseByDefault() {
+        val isLowEmphasis = layout.menuItemBase.isLowEmphasis
+
+        assertEquals(false, isLowEmphasis)
+    }
+
+    @Test
+    fun checksColorIsLowEmphasisWhenSetTrue() {
+        layout.menuItemBase.isLowEmphasis = true
+
+        assertEquals(getColorTokenFromTheme(activityController.get(), R.attr.colorLowEmphasis), layout.menuItemBase.getLabelColor())
+    }
+
+    @Test
+    fun checksColorIsLowEmphasisWhenSetFalse() {
+        layout.menuItemBase.isLowEmphasis = false
+
+        assertNotEquals(getColorTokenFromTheme(activityController.get(), R.attr.colorLowEmphasis), layout.menuItemBase.getLabelColor())
     }
 }